BBB-1158 Use BigDecimal for badge cost in pounds but send pence to gov pay

+++ database-schema/scripts/20190125141249_BBB-1158-create-payment-table.sql
@@ -7,7 +7,7 @@ CREATE TABLE payment.payment(
   payment_id                VARCHAR (150) NOT NULL,
   la_short_code             VARCHAR(10) NOT NULL,
   reference                 VARCHAR(100) NOT NULL,
-  cost                      INTEGER NOT NULL,
+  cost                      NUMERIC (10,2) NOT NULL,
   created_on                TIMESTAMP NOT NULL DEFAULT now(),
   PRIMARY KEY (payment_journey_uuid)
 );
